Test for high volume services table content
